+++ src/server/core_routes.js
@@ -1,14 +1,8 @@
 const express = require('express');
 const routes = express();
 
-routes.get('/test', (req, res) => {
-  res.status(200);
-  res.send('Hello World!');
-});
 
-routes.get('/', (req, res) => {
-  res.status(404).send('Nothing to see here 404');
-});
+
 
 routes.get('/api/user/:id', (req,res) =>{
   if(req.params.id == '1'){
