+++ src/server/server.js
@@ -1,8 +1,10 @@
-//const express = require('express');
-//const server = express();
+const express = require('express');
+const server = express();
 const morgan = require('morgan'); //HTTP Logger
 const config = require('config'); //Configuration helper
-const server = require("./core_routes"); 
+const baseRoutes = require("./base_routes"); 
+const coreRoutes = require("./core_routes"); 
+
 
 
 //this.startup();
@@ -17,9 +19,16 @@ const server = require("./core_routes");
   process.on('SIGINT', shutDown); 
   console.log('SIGTERM & SIGINT registered to shutdown gracefully...');  
 
+  //Add any global middle ware
   const morganFormat = config.get('expressSettings.morganFormat');
   server.use(morgan(morganFormat));
   console.log('Express HTTP request logger started with mode('+morganFormat+')!');  
+
+  //server.use(express.urlencoded({extended: false}));
+
+  //Add routes
+  server.use("/", baseRoutes);
+  server.use("/", coreRoutes);
 
   const port = config.get('expressSettings.serverPort');  
 
@@ -31,7 +40,6 @@ const server = require("./core_routes");
 //}
 
 
-console.log('SIGTERM & SIGINT registered to shutdown gracefully...');
 function shutDown() {
   console.log('Received kill signal, shutting down gracefully');
   //TODO:Cleanup here if needed
